Removed bug (in read function calls) in plot_rotation_curve.py and modified plot_Vmax_vs_V1kpc.py to allow switching between datasets easily and to not plot datapoints corresponding to haloes with less than 10 particles within 1kpc
The bug was due to hardcoded setting of the read function to read from
the med-res data. The bug was removed by changing the argument to these
functions from 'MR' to self.dataset. The program was plotting MR data
and therefore the "outlaying" haloes in the Vmax vs. V1kpc plot did not
correspond to these haloes plotted by this program.