Traced bug in rotation curve plot to the periodic wrap calculation and moved plot figures to their own directory
Commenting out the periodic wrapping of the coordinates made the
rotation curves look a lot more believable. To find out what was causing
the bug, I created the plot_halo.py script, which plots all particles
(regardless of type) of a given halo in a volume slice centered on the
COP of the halo.