Modified the satellite distribution by Vmax script to only plot satellites from one or the other of the central haloes
I also removed the possibility to plot isolated galaxies, since that is not very interesting to plot. In addition, the script may now plot the two LG-analogues instead of two different datasets in the same figure. This functionality is, however, very badly implemented at the moment and the flexibility and clarity of the code have worsened as an effect of this.