Modified rotation curve scripts to plot satellites in the appointed region
The plot_sat_rotation_profiles scripts was earlier plotting all satellites in a given halo. I worked a bit to modify to have the same function as the other satellite rotation profile plotting script, but be much more efficient. The script is not working at this moment -- I decided that the other working script is fast enough and therefore this script might be useless.