Dataset class now replaces the read_data functions for reading subhalos
The Dataset class now combines all the hdf5-files into a single File
object, which is constructed in the new function make_group_file, which
is called in __init__. The read_subhalo_attr function is basically
copied from the old module: it reads the given attribute values from the
data files. get_subhalos is the public interface, which uses
read_subhalo_attr. It allows for splitting the dataset into satellites
and isolated galaxies. The retrieval of the datasets might have been
slowed down, since other conditions (like SM > 0) are not imposed on the
datasets in these functions.