Created dataset_compute module
The new dataset_compute module contains functions for calculating new
quantities on a dataset. The calcVelocitiesAt1kpc, which was renamed to
calculate_V1kpc, was moved there from the Dataset class, along with the
get_subhalo_part_idx function. The purpose of the Dataset class is to
provide an interface to the data in a snapshot, and its extended
attributes. Functions that calculate the values of those extensions
should not be implemented in the Dataset class.