Match_snapshots looks into multiple files in the neighborhood of a halo
Added a function 'neighborhood' that is called by match_snapshots. One
never wants to look into all files but rather to the files near the one
that contains the reference halo. The number of files in a snapshot is
of the order of 100. If one sets the termination parameter to 100 halos,
then on average ~1-10 files are read. Thus, considerable improvement in
efficiency is achieved by this.