SOLVED: Found the bug resulting in weird rotation curves
In the method get_particles of the Snapshot class, I was first looping
through files and then, inside each file, through particle types. This
was exactly the opposite order from how the particles are iterated in
get_particle_masses. Thus, a particle could end up in different places
in the masses array and the groupnumber array, for example. Thus, I
switched the order of iteration in get_particles to: first loop through
part types and then through files. Since the particle file order does
not seem to have any meaning, this also seems like a more natural order
of iteration. This was also the root of my occational problems with V1kpc.