Corrected link ordering in part_file of Snapshot
The link numbers did not match with the corresponding file numbers,
which I saw as a potential future cause for a bug. After the correction,
the structure of the make_part_file function is identical to that of
make_group_file: the file names are read to an array and the sorted by
the file number to make sure that the numbers of the links, which are then
generated, match with them.