Rename functions and add docs in trace_halo
The purposes of the functions have changed from when they were named
first. The new functions also more clearly manifest the logic (if such
exists) of the tracing process: there is only one iteration, namely, the
iteration of the step; the index of the next tried subhalo is determined
by step, in a way that is implemented in get_index_at_step.

Renamings:  iteration -> get_index_at_step,
            get_matching_pair -> iterate_step