Add condition on masses of matched halos
The method for matching snapshots was updated and the requirements for a
match to be made now include a condition on the masses of the halos. In
addition, the return value type of the Snapshot.get_subhalos_IDs function
was changed from a dataset of lists to a dataset of ndarrays, as this is
more convenient for the matching method.