Create subhalo module for tracing individual subhalos
The new subhalo module replaces some methods from the trace_halo module.
The trace_halo module is responsible for finding matches and the Subhalo
class is meant to represent physical subhalos across snapshots.

The halo tracing functionality is tested with simple unit tests and some
notebooks.

On the side a get_redshift function is added to Snapshot.