Save traced halos and isolate tracing all halos into a new class
Tracing the halos takes a lot of time, but it only needs to be done
once, in principle. As the halo tracing methods are extended this way,
it makes sense to create a new, separate module for them. Furthermore,
a single halo can now be easily accessed in different snapshots with the
subhalo class.