SimulationTracer saves in hdf5 format instead of numpy
Hdf5 has the advantage that attributes can be added to datasets. Here,
saving the earliest traced snapshot helps a lot when reading the
tracers. The Subhalo class is also modified accordingly. The purpose of
this class is to represent a simple interface to the data associated
with a single halo. Thus, scripts for plotting a single halo are
significantly easier to follow. The tracing method of the Subhalo class
is removed, as all tracing is now expected to be done in bulk. Thus,
the Subhalo object takes as input the simulation tracer.

Note that the Snapshot class is also slightly modified: when reading
attributes, link0 is now used instead link1. This is because some
datasets might only have data in one file. It was assumed that link0
would work just as well as link1.