Save mass accumulation arrays to file when computed
The arrays are written under to "external" group in the combined group
data file. All the bound particles are written in one long array, so
that the saved object is transformable to an hdf5 dataset. Particles are
ordered first by halo, then by particle type, and lastly by distance to
host halo. The method mass_accumulation_to_array offers an interface to
this entry, transforming it to a more convenietly ordered arrays.