Add argument to get_subhalos in Snapshot to specify dataset HDF5 group
The method for creating the extension datasets, which are saved to the
"Extended" group of the HDF5 file, logically belongs to the
data_file_manipulation module, rather than dataset_compute, which is
merely for manipulating the datasets themselves. Previously, the actual
saving of the extension datasets was done in a Snapshot method. Now,
this is also the responsibility of the data_file_manipulation module.
Snapshot only reads the datafiles, and does not make modifications.
Instead, all modifications are are executed in the other module.