Modify and check subhalo dataset splitting methods
The methods now return mask arrays instead of splitting the actual data
arrays. This makes it easier to keep track of indices and the use of the
methods more flexible. There are two improvements introduced to these
methods in recent commits: the option to indicate MW and M31, and
another way of distinguishing satellites from isolated galaxies by
distance to central galaxies instead of group numbers. In this commit,
these improvements are implemented in a more reasonable way and also
tested with unit test methods as well as integrated tests in notebooks.