Order grouped particle datasets by particle type
When the particle data grouping by subhalo method is called, the
datasets for each subhalo are ordered by particle type. This is just
conventional -- one could always just call the method for each particle
type. But it is now easy to extend this method to other orderings if
such a need arises.

The mass accumulation computation method return value type is switched
from ndarray of ndarrays to list of ndarrays. This makes concatenation
more convenient.