Replace SimulationTracer with MergerTree
With the MergerTree class, all the subhalos through all the snapshots in
a simulation are traced either forward or backward in time. The matches
are saved in HDF5 format in the subhalo catalogues as group extensions.
Thus, all snapshots are treated equivalently. Instead, a new class that
uses the information contained in a MergerTree should be built to
represent the evolution of subhalos in a particular snapshot.