Complete the SnapshotTracer class
The class reads a MergerTree and traces subhalos in a given snapshot.
This class is now running as expected. The notebooks that make use of
this class, which plot the fraction of traced subhalos from a particular
snapshot are implemented.

Note that, contrary to its implementation in the previous commits, the
class assumes internally that the labeling for snapshots in any
simulation dataset starts at 0. This makes it easy to represent the
snapshots in an array, and increases compartmentalization in the package
- the SnapshotTracer class should not need to know about, which data
entries are actually available in the disk.