Create DataEnvelope class
This class implements and interface to the HDF5 file that envelopes all
data of a given snapshot. It is responsible for creating and
manipulating that file. The source data files are added as links to the
envelope file, and they are only read. DataEnvelope implements a method
that allows one to save additional datasets beside those links into the
envelope file.

Due to the particular organization of the datasets in the source files,
DataEnvelope does not, at this stage, implement methods for reading the
source files (since the access of many relevant data requires ad hoc
computations that do not fit well into the abstract purpose of this
class). Instead, the Snapshot object is still responsible for reading
them, but the access to the envelope file now takes place through the
group_data and particle_data (DataEnvelope objects) class attributes of
Snapshot.